requirements.txt
pyTelegramBotAPI==4.12.0
yt-dlp==2026.02.18
Flask==2.3.6
pycryptodome==3.19.0